+++ pkg/profiles/profiles.go
@@ -3,15 +3,19 @@ package profiles
 import (
 	"bytes"
 	"errors"
+	"fmt"
 	"io"
+	"os"
 	"text/template"
 	"time"
 
+	"github.com/ghodss/yaml"
 	"github.com/golang/protobuf/ptypes/duration"
 	pb "github.com/linkerd/linkerd2-proxy-api/go/destination"
 	sp "github.com/linkerd/linkerd2/controller/gen/apis/serviceprofile/v1alpha1"
 	"github.com/linkerd/linkerd2/pkg/util"
 	log "github.com/sirupsen/logrus"
+	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type profileTemplateConfig struct {
@@ -29,6 +33,11 @@ var (
 		Ttl: &duration.Duration{
 			Seconds: 10,
 		},
+	}
+	// ServiceProfileMeta is the TypeMeta for the ServiceProfile custom resource.
+	ServiceProfileMeta = meta_v1.TypeMeta{
+		APIVersion: "linkerd.io/v1alpha1",
+		Kind:       "ServiceProfile",
 	}
 	// DefaultServiceProfile is used for services with no service profile.
 	DefaultServiceProfile = pb.DestinationProfile{
@@ -408,5 +417,21 @@ func RenderProfileTemplate(namespace, service, controlPlaneNamespace string, w i
 	}
 
 	_, err = w.Write(buf.Bytes())
+	return err
+}
+
+func readFile(fileName string) (io.Reader, error) {
+	if fileName == "-" {
+		return os.Stdin, nil
+	}
+	return os.Open(fileName)
+}
+
+func writeProfile(profile sp.ServiceProfile, w io.Writer) error {
+	output, err := yaml.Marshal(profile)
+	if err != nil {
+		return fmt.Errorf("Error writing Service Profile: %s", err)
+	}
+	_, err = w.Write(output)
 	return err
 }